change genesis model path
**Phenomenon and reproduction steps**
merge genesis model to controller model
add genesis debug api

**Root cause and solution**

**Impactions**

**Test method**

**Affected branch(es)**

* main

**Checklist**

- [ ] Dependencies update required
- [ ] Common bug (similar problem in other repo)
